Add Delete Post

+++ controllers/feed.js
@@ -272,7 +272,39 @@ exports.updatePost = (req, res, next) => {
     });
 };
 
+exports.deletePost = (req,res,next) =>{
+  const postId = req.params.postId;
+  // mongoose function findById
+  Post.findById(postId).then(post =>{
+    if(!post){
+      const error = new Error("could not find post")
+      error.statusCode = 404;
+      throw error
+    }
+    clearImage(post.imageUrl) 
+
+    return Post.findByIdAndDelete(postId)
+
+  }).then(result => {
+    console.log(result)
+    res.status(200).json({message:"Deleted Post."})
+  }).catch(err =>{
+    if(!err.statusCode){
+      err.statusCode = 500;
+    }
+    next(err)
+  })
+}
+// const clearImage = filePath => {
+//   filePath = path.join(__dirname, '..', filePath);
+//   fs.unlink(filePath, err => console.log(err));
+// };
+
 const clearImage = filePath => {
   filePath = path.join(__dirname, '..', filePath);
-  fs.unlink(filePath, err => console.log(err));
+  fs.exists(filePath, exists => {
+    if (exists) {
+      fs.unlink(filePath, err => console.log(err));
+    }
+  });
 };
